Front_Back_delete_subscription_from_profile

+++ back/src/main/java/com/openclassrooms/mddapi/repositories/SubjectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubjectRepository extends CrudRepository<Subject, Integer> {
+public interface SubjectRepository extends CrudRepository<Subject, Long> {
 }
